add tests and refactor
