--- notebooks/reweave/reweave_utils.py
@@ -1,2 +1,0 @@
-from classes import *
-from reweave_utils import *
